+++ Morpher.WebApi/Ukrainian/UkrainianWebAnalyzer.cs
@@ -25,7 +25,7 @@
             }
         }
 
-        public NumberSpelling Spell(int n, string unit)
+        public NumberSpelling Spell(decimal n, string unit)
         {
             try
             {
